+++ test/utils/translate.test.ts
@@ -16,9 +16,11 @@ jest.mock('../../src/services/languageService', () => ({
     settingsHeader: 'Settings',
     languageHeader: 'Languages',
     switch2DButton: '2D Mode',
+    switchFlatButton: 'Flat Mode',
     switchWheelchairModeButton: 'Wheelchair Mode',
     zoomInButton: 'Zoom in',
     zoomOutButton: 'Zoom out',
+    centeringButton: 'Center map',
     changeLevel: 'Change to level ',
     showPreviousLevels: 'Show previous levels',
     showNextLevels: 'Show next levels',
@@ -34,10 +36,11 @@ describe('translate()', () => {
       <ul id="userProfileList"></ul>
       <ul id="userSettingsList"></ul>
       <ul id="languageList"></ul>
-      <button id="switch2DLabel"></button>
+      <button id="switch2D"><span id="switch2DLabel"></span></button>
       <button id="switchWheelchairMode"></button>
-      <button id="zoomControlInLabel"></button>
-      <button id="zoomControlOutLabel"></button>
+      <button id="zoomControlIn"><span id="zoomControlInLabel"></span></button>
+      <button id="zoomControlOut"><span id="zoomControlOutLabel"></span></button>
+      <button id="centeringButton"><span aria-hidden="true"></span></button>
       <button id="levelShiftUp"></button>
       <button id="levelShiftDown"></button>
       <ul id="levelControl">
@@ -56,16 +59,19 @@ describe('translate()', () => {
   it('updates aria-labels and titles', () => {
     translate();
     const userProfileList = getRequiredElement('userProfileList');
-    const switch2DLabel = getRequiredElement('switch2DLabel');
-    const zoomControlInLabel = getRequiredElement('zoomControlInLabel');
+    const switch2D = getRequiredElement('switch2D');
+    const zoomControlIn = getRequiredElement('zoomControlIn');
+    const centeringButton = getRequiredElement('centeringButton');
     const levelShiftUp = getRequiredElement('levelShiftUp');
     const levelButton = getRequiredElement('levelControl').children[0].firstElementChild as HTMLElement;
 
     expect(userProfileList.ariaLabel).toBe('Profiles');
-    expect(switch2DLabel.title).toBe('2D Mode');
-    expect(switch2DLabel.ariaLabel).toBe('2D Mode');
-    expect(zoomControlInLabel.title).toBe('Zoom in');
-    expect(zoomControlInLabel.ariaLabel).toBe('Zoom in');
+    expect(switch2D.title).toBe('2D Mode');
+    expect(switch2D.ariaLabel).toBe('2D Mode');
+    expect(zoomControlIn.title).toBe('Zoom in');
+    expect(zoomControlIn.ariaLabel).toBe('Zoom in');
+    expect(centeringButton.title).toBe('Center map');
+    expect(centeringButton.ariaLabel).toBe('Center map');
     expect(levelShiftUp.title).toBe('Show previous levels');
     expect(levelShiftUp.ariaLabel).toBe('Show previous levels');
     expect(levelButton.title).toBe('Change to level 1');